requirements.txt
langchain>=0.1.0
langchain-ollama>=0.0.1
langgraph>=0.0.20
langchain-community>=0.0.13
pydantic>=2.0.0
python-dotenv>=1.0.0 